requirements.txt
asgiref==3.9.1
Django==4.2.23
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
phonenumbers==8.13.55
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.3
typing_extensions==4.14.1
